Implement Secure version of BCryptPasswordEncoder
Currently all implementations of spring's `PasswordEncoder`, takes in `String`,
or does `.toString`. This leads of a leak in the password in the memory,
chance of interned and leak. This implementation takes care by creating
a wrapper on char[] for accepting password, a Secure version of
BCryptPasswordEncoder, which never converts to string, and cleans up the
backing buffers respectively.